Make sure CIVICRM_CLEANURL is always on for when using cv
Defines the constant in drushrc.php, and then makes sure that this file
is included from civicrm.settings.php, instead of local.drushrc.php.

We were having issues with using cv for cron/monitoring, because of how
it bootstraps differently and CIVICRM_CLEANURL was not defined in time
for the monitoring check, which also led to weird permalink caching issues.